Code refactoring for lambda
