+++ .example_env
@@ -1,3 +1,4 @@
 PORT = 5432
 LLM_MODEL = "anthropic.claude-3-sonnet-20240229-v1:0"
 EMBEDDING_MODEL = "amazon.titan-embed-text-v2:0"
+ORCHESTRATOR_LAMBDA_FUNCTION_NAME = "orchastrator_lambda_function_name"
