+++ lambda_functions/classification_node/__init__.py
@@ -1,1 +1,1 @@
-from .classification import lambda_handler
+from .main import lambda_handler
