+++ lambda_functions/classification_node/main.py
