+++ lambda_functions/embedding_tool/__init__.py
@@ -1,1 +1,1 @@
-from .embedding import lambda_handler
+from .main import lambda_handler
