+++ lambda_functions/letter_res_node/main.py
