+++ lambda_functions/orchestrator/__init__.py
@@ -1,1 +1,1 @@
-from .orchestrator import lambda_handler
+from .main import lambda_handler
