+++ lambda_functions/response_node/__init__.py
@@ -1,1 +1,1 @@
-from .response import lambda_handler
+from .main import lambda_handler
